Core/Movement: Fixed data sent in cyclic spline related packets
Port From (https://github.com/TrinityCore/TrinityCore/commit/a28facb95633435f89f134eec55a4a545251be9b)

+++ Source/Game/Entities/Unit/Unit.Movement.cs
@@ -1818,7 +1818,7 @@ namespace Game.Entities
 
                     FlightSplineSync flightSplineSync = new();
                     flightSplineSync.Guid = GetGUID();
-                    flightSplineSync.SplineDist = MoveSpline.TimePassed() / MoveSpline.Duration();
+                    flightSplineSync.SplineDist = (float)MoveSpline.TimePassed() / MoveSpline.Duration();
                     SendMessageToSet(flightSplineSync, true);
                 }
             }
